+++ bin/commands/inbox/select.rs
@@ -24,7 +24,6 @@ use radicle_tui as tui;
 
 use tui::store;
 use tui::store::StateValue;
-use tui::ui::items::{Filter, NotificationItem, NotificationItemFilter};
 use tui::ui::span;
 use tui::ui::widget::container::{Column, Container, Footer, FooterProps, Header, HeaderProps};
 use tui::ui::widget::text::{TextArea, TextAreaProps};
@@ -33,6 +32,7 @@ use tui::ui::widget::{ToWidget, Widget};
 use tui::{BoxedAny, Channel, Exit, PageStack};
 
 use crate::cob::inbox;
+use crate::ui::items::{Filter, NotificationItem, NotificationItemFilter};
 
 use self::ui::Browser;
 use self::ui::BrowserProps;