+++ bin/commands/issue/select.rs
@@ -19,7 +19,6 @@ use radicle_tui as tui;
 
 use tui::store;
 use tui::store::StateValue;
-use tui::ui::items::{Filter, IssueItem, IssueItemFilter};
 use tui::ui::span;
 use tui::ui::widget::container::{Column, Container, Footer, FooterProps, Header, HeaderProps};
 use tui::ui::widget::text::{TextArea, TextAreaProps};
@@ -29,6 +28,7 @@ use tui::ui::widget::{ToWidget, Widget};
 use tui::{BoxedAny, Channel, Exit, PageStack};
 
 use crate::cob::issue;
+use crate::ui::items::{Filter, IssueItem, IssueItemFilter};
 
 use self::ui::{Browser, BrowserProps};