+++ src/ui.rs
@@ -1,6 +1,4 @@
 pub mod ext;
-pub mod format;
-pub mod items;
 pub mod layout;
 pub mod span;
 pub mod theme;