+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Move all Radicle-dependent code to `bin/`
+
 ## [0.3.1] - 2024-06-11
 
 ### Added